requirements.txt
pytest
pydantic
python-dotenv
arcgis